+++ lib/aws-sessionstore-dynamodb.rb
@@ -1,19 +1,19 @@
+# frozen_string_literal: true
+
 module Aws
   module SessionStore
-    module DynamoDB; end
+    # Namespace for DynamoDB rack session storage.
+    module DynamoDB
+      VERSION = File.read(File.expand_path('../VERSION', __dir__)).strip
+    end
   end
 end
 
-require 'aws/session_store/dynamo_db/configuration'
-require 'aws/session_store/dynamo_db/invalid_id_error'
-require 'aws/session_store/dynamo_db/missing_secret_key_error'
-require 'aws/session_store/dynamo_db/lock_wait_timeout_error'
-require 'aws/session_store/dynamo_db/errors/base_handler'
-require 'aws/session_store/dynamo_db/errors/default_handler'
-require 'aws/session_store/dynamo_db/garbage_collection'
-require 'aws/session_store/dynamo_db/locking/base'
-require 'aws/session_store/dynamo_db/locking/null'
-require 'aws/session_store/dynamo_db/locking/pessimistic'
-require 'aws/session_store/dynamo_db/rack_middleware'
-require 'aws/session_store/dynamo_db/table'
-require 'aws/session_store/dynamo_db/version'
+require_relative 'aws/session_store/dynamo_db/configuration'
+require_relative 'aws/session_store/dynamo_db/errors'
+require_relative 'aws/session_store/dynamo_db/garbage_collection'
+require_relative 'aws/session_store/dynamo_db/locking/base'
+require_relative 'aws/session_store/dynamo_db/locking/null'
+require_relative 'aws/session_store/dynamo_db/locking/pessimistic'
+require_relative 'aws/session_store/dynamo_db/rack_middleware'
+require_relative 'aws/session_store/dynamo_db/table'