+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Unreleased Changes
 ------------------
 
+* Feature - Uses `rack` version 3 as the minimum.
+
+* Feature - Drop support for Ruby 2.5 and 2.6.
+
+* Feature - Support additional configuration options through ENV.
+
+* Feature - Moves Error classes into the Errors module.
+
 2.2.0 (2024-01-25)
 ------------------